Devl

+++ src/main/java/com/fseteam1/amazingcomputerparts/service/ComputerPartServiceIMPL.java
@@ -1,6 +1,6 @@
 /*
   FSE Cohort - Team Project2
-  Amazing Computer Parts Inventory System -Team2
+  Amazing Computer Parts Inventory System -Team1
   Class: ComputerPartServiceIMPL - service implementation
  */
 
@@ -8,6 +8,8 @@ package com.fseteam1.amazingcomputerparts.service;
 
 import com.fseteam1.amazingcomputerparts.dao.IComputerPartDAO;
 import com.fseteam1.amazingcomputerparts.entity.ComputerPart;
+import com.fseteam1.amazingcomputerparts.exception.PartNotFoundForDeleteException;
+import com.fseteam1.amazingcomputerparts.exception.PartNotFoundForGetException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
@@ -31,7 +33,11 @@ public class ComputerPartServiceIMPL implements IComputerPartService{
 
     @Override
     public ComputerPart findById(int computerPartId) {
-        return computerPartDAO.findById(computerPartId);
+        if (computerPartDAO.findById(computerPartId) == null) {
+            throw new PartNotFoundForGetException(computerPartId);
+        } else {
+            return computerPartDAO.findById(computerPartId);
+        }
     }
 
     @Override
@@ -41,6 +47,10 @@ public class ComputerPartServiceIMPL implements IComputerPartService{
 
     @Override
     public void deleteById(int computerPartId) {
-        computerPartDAO.deleteById(computerPartId);
+        if (computerPartDAO.findById(computerPartId) == null) {
+            throw new PartNotFoundForDeleteException(computerPartId);
+        } else {
+            computerPartDAO.deleteById(computerPartId);
+        }
     }
 }
